requirements.txt
click==8.0.1
colorama==0.4.4
futures==3.0.5
google-auth==2.17.3
google-auth-oauthlib==0.4.5
google-auth-httplib2==0.1.0
google-api-python-client==2.22.0
google-cloud-logging==2.7.0
google-cloud-compute==1.11.0
emoji==1.6.1
jinja2==3.0.1
requests==2.26.0
urllib3==1.26.6
chardet==4.0.0
pyinstaller==5.10.0
pillow==9.5.0
google-cloud-storage==2.8.0
google-cloud-container==2.21.0